# examples/petstore/api/api_pet.go
/*
 * Swagger Petstore
 *
 * This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.
 *
 * API version: 1.0.3
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A PetApiController binds http requests to an api service and writes the service results to the http response
type PetApiController struct {
	service PetApiServicer
}

// NewPetApiController creates a default api controller
func NewPetApiController(s PetApiServicer) Router {
	return &PetApiController{service: s}
}

// Routes returns all of the api route for the PetApiController
func (c *PetApiController) Routes() Routes {
	return Routes{
		{
			"AddPet",
			strings.ToUpper("Post"),
			"/v2/pet",
			c.AddPet,
		},
		{
			"DeletePet",
			strings.ToUpper("Delete"),
			"/v2/pet/{petId}",
			c.DeletePet,
		},
		{
			"FindPetsByStatus",
			strings.ToUpper("Get"),
			"/v2/pet/findByStatus",
			c.FindPetsByStatus,
		},
		{
			"FindPetsByTags",
			strings.ToUpper("Get"),
			"/v2/pet/findByTags",
			c.FindPetsByTags,
		},
		{
			"GetPetById",
			strings.ToUpper("Get"),
			"/v2/pet/{petId}",
			c.GetPetById,
		},
		{
			"UpdatePet",
			strings.ToUpper("Put"),
			"/v2/pet",
			c.UpdatePet,
		},
		{
			"UpdatePetWithForm",
			strings.ToUpper("Post"),
			"/v2/pet/{petId}",
			c.UpdatePetWithForm,
		},
		{
			"UploadFile",
			strings.ToUpper("Post"),
			"/v2/pet/{petId}/uploadImage",
			c.UploadFile,
		},
	}
}

// AddPet - Add a new pet to the store
func (c *PetApiController) AddPet(w http.ResponseWriter, r *http.Request) {
	body := &Pet{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&body); err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var serviceErr error
	serviceErr = c.service.AddPet(r.Context(), *body)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// DeletePet - Deletes a pet
func (c *PetApiController) DeletePet(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	petId, err := parseIntParameter(params["petId"])
	if err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	apiKey := r.Header.Get("apiKey")

	var serviceErr error
	serviceErr = c.service.DeletePet(r.Context(), petId, apiKey)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// FindPetsByStatus - Finds Pets by status
func (c *PetApiController) FindPetsByStatus(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	status := strings.Split(query.Get("status"), ",")
	var result []Pet
	var serviceErr error
	result, serviceErr = c.service.FindPetsByStatus(r.Context(), status)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}

	EncodeJSONResponse(result, nil, w)
}

// FindPetsByTags - Finds Pets by tags
func (c *PetApiController) FindPetsByTags(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	tags := strings.Split(query.Get("tags"), ",")
	var result []Pet
	var serviceErr error
	result, serviceErr = c.service.FindPetsByTags(r.Context(), tags)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}

	EncodeJSONResponse(result, nil, w)
}

// GetPetById - Find pet by ID
func (c *PetApiController) GetPetById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	petId, err := parseIntParameter(params["petId"])
	if err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var result Pet
	var serviceErr error
	result, serviceErr = c.service.GetPetById(r.Context(), petId)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}

	EncodeJSONResponse(result, nil, w)
}

// UpdatePet - Update an existing pet
func (c *PetApiController) UpdatePet(w http.ResponseWriter, r *http.Request) {
	body := &Pet{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&body); err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var serviceErr error
	serviceErr = c.service.UpdatePet(r.Context(), *body)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// UpdatePetWithForm - Updates a pet in the store with form data
func (c *PetApiController) UpdatePetWithForm(w http.ResponseWriter, r *http.Request) {
	err := r.ParseForm()
	if err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	params := mux.Vars(r)
	petId, err := parseIntParameter(params["petId"])
	if err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	name := r.FormValue("name")
	status := r.FormValue("status")

	var serviceErr error
	serviceErr = c.service.UpdatePetWithForm(r.Context(), petId, name, status)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// UploadFile - uploads an image
func (c *PetApiController) UploadFile(w http.ResponseWriter, r *http.Request) {
	err := r.ParseForm()
	if err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	params := mux.Vars(r)
	petId, err := parseIntParameter(params["petId"])
	if err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	additionalMetadata := r.FormValue("additionalMetadata")
	file, err := ReadFormFileToTempFile(r, "file")
	if err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var result ApiResponse
	var serviceErr error
	result, serviceErr = c.service.UploadFile(r.Context(), petId, additionalMetadata, file)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}

	EncodeJSONResponse(result, nil, w)
}