# examples/petstore/api/api_user.go
/*
 * Swagger Petstore
 *
 * This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.
 *
 * API version: 1.0.3
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A UserApiController binds http requests to an api service and writes the service results to the http response
type UserApiController struct {
	service UserApiServicer
}

// NewUserApiController creates a default api controller
func NewUserApiController(s UserApiServicer) Router {
	return &UserApiController{service: s}
}

// Routes returns all of the api route for the UserApiController
func (c *UserApiController) Routes() Routes {
	return Routes{
		{
			"CreateUser",
			strings.ToUpper("Post"),
			"/v2/user",
			c.CreateUser,
		},
		{
			"CreateUsersWithArrayInput",
			strings.ToUpper("Post"),
			"/v2/user/createWithArray",
			c.CreateUsersWithArrayInput,
		},
		{
			"CreateUsersWithListInput",
			strings.ToUpper("Post"),
			"/v2/user/createWithList",
			c.CreateUsersWithListInput,
		},
		{
			"DeleteUser",
			strings.ToUpper("Delete"),
			"/v2/user/{username}",
			c.DeleteUser,
		},
		{
			"GetUserByName",
			strings.ToUpper("Get"),
			"/v2/user/{username}",
			c.GetUserByName,
		},
		{
			"LoginUser",
			strings.ToUpper("Get"),
			"/v2/user/login",
			c.LoginUser,
		},
		{
			"LogoutUser",
			strings.ToUpper("Get"),
			"/v2/user/logout",
			c.LogoutUser,
		},
		{
			"UpdateUser",
			strings.ToUpper("Put"),
			"/v2/user/{username}",
			c.UpdateUser,
		},
	}
}

// CreateUser - Create user
func (c *UserApiController) CreateUser(w http.ResponseWriter, r *http.Request) {
	body := &User{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&body); err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var serviceErr error
	serviceErr = c.service.CreateUser(r.Context(), *body)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// CreateUsersWithArrayInput - Creates list of users with given input array
func (c *UserApiController) CreateUsersWithArrayInput(w http.ResponseWriter, r *http.Request) {
	body := &[]User{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&body); err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var serviceErr error
	serviceErr = c.service.CreateUsersWithArrayInput(r.Context(), *body)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// CreateUsersWithListInput - Creates list of users with given input array
func (c *UserApiController) CreateUsersWithListInput(w http.ResponseWriter, r *http.Request) {
	body := &[]User{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&body); err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var serviceErr error
	serviceErr = c.service.CreateUsersWithListInput(r.Context(), *body)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// DeleteUser - Delete user
func (c *UserApiController) DeleteUser(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	username := params["username"]

	var serviceErr error
	serviceErr = c.service.DeleteUser(r.Context(), username)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// GetUserByName - Get user by user name
func (c *UserApiController) GetUserByName(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	username := params["username"]
	var result User
	var serviceErr error
	result, serviceErr = c.service.GetUserByName(r.Context(), username)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}

	EncodeJSONResponse(result, nil, w)
}

// LoginUser - Logs user into the system
func (c *UserApiController) LoginUser(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	username := query.Get("username")
	password := query.Get("password")
	var result string
	var serviceErr error
	result, serviceErr = c.service.LoginUser(r.Context(), username, password)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}

	EncodeJSONResponse(result, nil, w)
}

// LogoutUser - Logs out current logged in user session
func (c *UserApiController) LogoutUser(w http.ResponseWriter, r *http.Request) {

	var serviceErr error
	serviceErr = c.service.LogoutUser(r.Context())
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}

// UpdateUser - Updated user
func (c *UserApiController) UpdateUser(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	username := params["username"]
	body := &User{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&body); err != nil {
		apiLog.Error(err)
		w.WriteHeader(500)
		return
	}

	var serviceErr error
	serviceErr = c.service.UpdateUser(r.Context(), username, *body)
	if serviceErr != nil {
		var ae *Error
		if errors.As(serviceErr, &ae) {
			w.WriteHeader(ae.Code)
			fmt.Fprint(w, ae.Error())
		} else {
			w.WriteHeader(500)
		}
		return
	}
}